feat: add depth channel extraction for HEIC images
- New module: shared_utils/src/depth_channel.rs
  - extract_depth_from_heic(): Extract depth maps from HEIC auxiliary images
  - encode_jxl_with_depth(): Encode JXL with depth support via jpegxl-rs
  - encode_jxl_depth_fallback(): Sidecar approach for depth storage
  - DepthMap and DepthType data structures

- HDR synthesis improvements (hdr_synthesis.rs)
  - Simultaneous processing of gainmap and depth auxiliary images
  - Depth map saved as sidecar file (.depth.png)
  - XMP metadata parsing for near/far focus distance

- Integration updates
  - image_analyzer.rs: Updated logging for depth/gainmap detection
  - image_formats.rs: Fixed WebP quality estimation integer overflow
  - lib.rs: Export depth_channel module

- Dependencies
  - Added jpegxl-rs v0.12 with vendored libjxl

- Tests
  - 4 unit tests for depth_channel module (all passing)

Note: Full HEIC gainmap/depth testing requires real device photos
(iPhone HDR, Samsung Super HDR, Google UltraHDR)

+++ shared_utils/src/image_formats.rs
@@ -299,7 +299,7 @@ pub mod webp {
                 let vp8_data = &data[payload_start..payload_end];
                 if vp8_data.len() >= 10 && vp8_data[3..6] == [0x9D, 0x01, 0x2A] {
                     let y_ac_qi = vp8_data[10] & 0x7F;
-                    let quality = ((127 - y_ac_qi) * 100 / 127).min(100);
+                    let quality = ((127 - y_ac_qi) as u32 * 100 / 127).min(100) as u8;
                     return Ok(quality);
                 }
             }
